colexec: miscellaneous cleanups
This commit does the following:
1. plumbs the processor ID of the processor that we're wrapping into the
vectorized flow. This processor ID will be used for the materializer
/ columnarizer pair. I don't think it matters much, but why not.
2. resolves a minor TODO in the external sorter benchmark.
3. removes limitation of `2<<16-1` on K in the top K sorter (previously
limited by `uint16` batch length.
4. removes calling `Next` on the input of `zeroOperator` (which always
returns a zero batch). I believe previously it was needed so that
projecting operators had a chance to append their output columns to the
batch which is no longer the case.

Release note: None

+++ pkg/sql/colexec/utils_test.go
@@ -1465,7 +1465,7 @@ type sortTestCase struct {
 	typs        []types.T
 	ordCols     []execinfrapb.Ordering_Column
 	matchLen    int
-	k           uint16
+	k           int
 }
 
 // Mock typing context for the typechecker.
